+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - README: Added Brink features overview to the introduction
 - INTERNAL: Enormous simplification of const eval to just use the AST
 - INTERNAL: Type deduction moves from IRDB to the Linearizer
+- INTERNAL: Replace `unsafe` mmap with OutputBuffer wrapper for vec<>
 
 ## [7.0.0] - 2026-05-02